+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 * `vm/vmss create`: support cross tenant image referencing.
 * `vm diagnostics get-default-config --windows-os`: fix bug in windows os default config.
 * `vmss extension set`: expose `--provision-after-extensions`. It defines what extensions must be provisioned before the extension being set.
+* `sig image-version update`: expose --replica-count for setting the default replication count.
 
 2.2.11
 ++++++